Move #glob_include? from core_ext to entry_filter
