+++ test/test_core_ext.rb
@@ -63,32 +63,4 @@ class TestCoreExt < Test::Unit::TestCase
     end
 
   end
-
-  context "enumerable" do
-    context "glob_include?" do
-      should "return false with no glob patterns" do
-        assert ![].glob_include?("a.txt")
-      end
-
-      should "return false with all not match path" do
-        data = ["a*", "b?"]
-        assert !data.glob_include?("ca.txt")
-        assert !data.glob_include?("ba.txt")
-      end
-
-      should "return true with match path" do
-        data = ["a*", "b?", "**/a*"]
-        assert data.glob_include?("a.txt")
-        assert data.glob_include?("ba")
-        assert data.glob_include?("c/a/a.txt")
-        assert data.glob_include?("c/a/b/a.txt")
-      end
-
-      should "match even if there is no leading slash" do
-        data = ['vendor/bundle']
-        assert data.glob_include?('/vendor/bundle')
-        assert data.glob_include?('vendor/bundle')
-      end
-    end
-  end
 end
